int64

+++ burbuja.go
@@ -1,8 +1,8 @@
 package main
 
 
-func burbuja(Lista []int){
- var auxiliar int
+func burbuja(Lista []int64){
+ var auxiliar int64
  for i := 0; i < len(Lista); i++ {
   for j := 0; j < len(Lista); j++ {
    if Lista[i] < Lista[j] {
